:art: re-organize solver structure
